+++ public/dialogue/dialogue.yarn
@@ -1,8 +1,13 @@
 title: Start
 ---
-Eva: what's up, [wave]homosexuals?[/wave] #emote:flushed
-Eva: have u ever thought deeply about [shake]veggie corndogs[/shake] #emote:thinking
-Sam: big brain scratcher #emote:flushed
-Eva: [shake]EVERYTHING IS OK[/shake] #emote:determined
-Eva: smiley
+Eva: what's up, homosexuals?
+-> Um.
+   Eva: Erm.
+-> Hello.
+   Eva: Hello.
+Eva: have u ever thought deeply about... #emote:thinking
+Eva: [shake]veggie corndogs[/shake] #emote:flushed
+Sam: big brain scratcher #emote:unamused
+Eva: [wave]STILL EATIN' EM![/wave] #emote:determined #big
+Eva: ok bye
 ===
