+++ src/components/dialogue/DialogueLineFragment.tsx
@@ -1,14 +1,49 @@
-import React, { Fragment } from "react";
-import { TypedCharacter } from "./TypedCharacter";
+import React, { Fragment, PropsWithChildren } from "react";
 import { nanoid } from "nanoid";
-import { EmphasizedCharacter } from "./EmphasizedCharacter";
 import styles from "./dialogue.module.css";
+import { Variants, motion } from "framer-motion";
 
-export enum FragmentAnimationStyle {
+export enum FragmentAnimationVariant {
   Wave = "wave",
   Shake = "shake",
   None = "none",
 }
+
+interface TypedCharacterProps extends PropsWithChildren {
+  /**
+   * Index of the character to be rendered
+   * to stagger animations
+   */
+  index: number;
+
+  /**
+   * Speed to display characters
+   * @default 0.04
+   */
+  speed?: number;
+
+  /**
+   * Amount to delay the initial typing sequence
+   * @default 0.2
+   */
+  delay?: number;
+}
+
+const TypedCharacter = ({
+  children,
+  index,
+  speed = 0.04,
+  delay = 0.2,
+}: TypedCharacterProps) => {
+  return (
+    <motion.span
+      className={styles.character}
+      style={{ animationDelay: `${index * speed + delay}s` }}
+    >
+      {children}
+    </motion.span>
+  );
+};
 
 export interface DialogueLineFragmentProps {
   /**
@@ -20,7 +55,7 @@ export interface DialogueLineFragmentProps {
    * Animation style of the sentence fragment
    * @default "none"
    */
-  style?: FragmentAnimationStyle;
+  variant?: FragmentAnimationVariant;
 
   /**
    * Starting character index for continuous animation
@@ -30,37 +65,138 @@ export interface DialogueLineFragmentProps {
   index?: number;
 }
 
+interface EmphasizedCharacterProps extends PropsWithChildren {
+  /**
+   * Index of the character to be rendered
+   * to stagger animations
+   */
+  index: number;
+
+  /**
+   * Animation style of the character
+   * @default "none"
+   */
+  variant?: FragmentAnimationVariant;
+}
+
+const EmphasizedCharacter = ({
+  children,
+  variant = FragmentAnimationVariant.None,
+  index,
+}: EmphasizedCharacterProps) => {
+  const generateRandomValuesArray = (
+    num: number,
+    min: number,
+    max: number
+  ): number[] => {
+    const randomNumberBetweenValues = (): number => {
+      return Math.random() * (max - min) + min;
+    };
+
+    return [...Array(num)].map(randomNumberBetweenValues);
+  };
+
+  const generateRandomEmsArray = (
+    num: number,
+    min: number,
+    max: number
+  ): string[] => {
+    return generateRandomValuesArray(num, min, max).map((i) => i + "em");
+  };
+
+  const shakeVariants: Variants = {
+    initial: (i) => ({
+      x: 0,
+      // Alternate heights
+      y: i % 2 === 0 ? "0.03em" : "-0.03em",
+      // Alternate rotation
+      rotate: i % 2 === 0 ? 5 : -5,
+    }),
+    animate: (i) => ({
+      x: generateRandomEmsArray(10, -0.05, 0.05),
+      y: generateRandomEmsArray(10, -0.1, 0.1),
+      rotate: generateRandomValuesArray(10, 10, -10),
+      transition: {
+        type: "tween",
+        repeat: Infinity,
+        repeatType: "mirror",
+        ease: "anticipate",
+        delay: -i,
+        duration: 1,
+      },
+    }),
+  };
+
+  const waveVariants: Variants = {
+    initial: { y: "-0.12em" },
+    animate: (i) => ({
+      y: "0.12em",
+      transition: {
+        type: "tween",
+        repeat: Infinity,
+        ease: "easeInOut",
+        repeatType: "reverse",
+        delay: -i * 0.05,
+        duration: 0.8,
+      },
+    }),
+  };
+
+  const getVariantObject = (variant?: FragmentAnimationVariant) => {
+    switch (variant) {
+      case FragmentAnimationVariant.Shake:
+        return shakeVariants;
+      case FragmentAnimationVariant.Wave:
+        return waveVariants;
+      default:
+        return undefined;
+    }
+  };
+
+  return (
+    <motion.strong
+      className={styles.emphasizedCharacter}
+      variants={getVariantObject(variant)}
+      initial="initial"
+      animate="animate"
+      custom={index}
+    >
+      {children}
+    </motion.strong>
+  );
+};
+
 export const DialogueLineFragment = ({
   text,
-  style,
+  variant,
   index = 0,
 }: DialogueLineFragmentProps) => {
   const words = text.split(" ");
 
-  const renderWordsAndChars = () => {
-    return words.map((word, wordIndex) => (
-      <span className={styles.word} key={nanoid()}>
-        {word.split("").map((char) => {
-          return (
-            <TypedCharacter key={nanoid()} index={index++}>
-              {style === undefined || style === FragmentAnimationStyle.None ? (
-                <>{char}</>
-              ) : (
-                <EmphasizedCharacter style={style} index={index}>
-                  {char}
-                </EmphasizedCharacter>
-              )}
-            </TypedCharacter>
-          );
-        })}
-        {wordIndex < words.length - 1 && (
-          <TypedCharacter key={nanoid()} index={index++}>
-            {" "}
-          </TypedCharacter>
-        )}
-      </span>
-    ));
-  };
+  const isEmphasized = !(
+    variant === undefined || variant === FragmentAnimationVariant.None
+  );
 
-  return renderWordsAndChars();
+  return words.map((word, wordIndex) => (
+    <span className={styles.word} key={nanoid()}>
+      {word.split("").map((char) => {
+        return (
+          <TypedCharacter key={nanoid()} index={index++}>
+            {isEmphasized ? (
+              <EmphasizedCharacter variant={variant} index={index}>
+                {char}
+              </EmphasizedCharacter>
+            ) : (
+              <>{char}</>
+            )}
+          </TypedCharacter>
+        );
+      })}
+      {wordIndex < words.length - 1 && (
+        <TypedCharacter key={nanoid()} index={index++}>
+          {" "}
+        </TypedCharacter>
+      )}
+    </span>
+  ));
 };
